+++ src/main/kotlin/me/elabee/idea/drupal/template/DrupalTwigCompletionProvider.kt
@@ -3,10 +3,13 @@ package me.elabee.idea.drupal.template
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.lookup.LookupElementPresentation
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.ModificationTracker
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiManager
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider
@@ -15,6 +18,7 @@ import com.intellij.util.ProcessingContext
 import com.intellij.util.Processor
 import com.intellij.util.indexing.FileBasedIndex
 import fr.adrienbrault.idea.symfony2plugin.templating.TemplateLookupElement
+import me.elabee.idea.drupal.getIcon
 import me.elabee.idea.drupal.indexing.DrupalIndexIds
 
 class DrupalTwigCompletionProvider : CompletionProvider<CompletionParameters>() {
@@ -32,6 +36,7 @@ class DrupalTwigCompletionProvider : CompletionProvider<CompletionParameters>() 
 
     private fun getLookupElementsFromIndex(project: Project): Set<TemplateLookupElement> {
         val projectDir = project.guessProjectDir() ?: return emptySet()
+        val psiManager = PsiManager.getInstance(project)
 
         return CachedValuesManager.getManager(project).getCachedValue(
             project,
@@ -49,8 +54,7 @@ class DrupalTwigCompletionProvider : CompletionProvider<CompletionParameters>() 
                             GlobalSearchScope.allScope(project),
                             null,
                             Processor {
-                                val twig = it.parent.findChild("${it.parent.name}.twig") ?: return@Processor true
-                                elements.add(TemplateLookupElement(name, twig, projectDir, true))
+                                elements.add(SdcLookupElement(name, it.parent, projectDir, psiManager))
                                 true
                             },
                         )
@@ -66,5 +70,25 @@ class DrupalTwigCompletionProvider : CompletionProvider<CompletionParameters>() 
             },
             false,
         )
+    }
+
+    private class SdcLookupElement : TemplateLookupElement {
+        private var psiManager: PsiManager
+        private var virtualFile: VirtualFile
+
+        constructor(templateName: String, virtualFile: VirtualFile, projectBaseDir: VirtualFile, psiManager: PsiManager) : super(
+            templateName,
+            virtualFile,
+            projectBaseDir,
+            true,
+        ) {
+            this.psiManager = psiManager
+            this.virtualFile = virtualFile
+        }
+
+        override fun renderElement(presentation: LookupElementPresentation) {
+            super.renderElement(presentation)
+            presentation.icon = virtualFile.getIcon(psiManager)
+        }
     }
 }